added logout

+++ utilities.go
@@ -93,9 +93,20 @@ func checkTokenValidity(token string) (valid bool, login string, err error) {
 		removal_err := os.Remove(fmt.Sprintf("./save_data/account_tokens/%s.json", token))
 		if removal_err != nil {
 			serverLog("utilities.go:95", fmt.Sprintf("failed to remove the token file \"%s.json\", the error \"%s\"", token, removal_err))
-			return false, "", nil
+			return false, "", removal_err
 		}
 		return false, "", nil
+	}
+
+	account_token_json, account_token_json_err := json.Marshal(AccountToken{token_data.Login, time.Now().Unix() + 1200})
+	if account_token_json_err != nil {
+		serverLog("utilities.go:103", fmt.Sprintf("could not marshal the account token data to json, the error \"%s\"", account_token_json_err))
+		return
+	}
+	account_token_file_err := os.WriteFile(fmt.Sprintf("./save_data/account_tokens/%s.json", token), account_token_json, 0644)
+	if account_token_file_err != nil {
+		serverLog("utilities.go:108", fmt.Sprintf("could not write the account token file \"%s.json\", the error \"%s\"", token, account_token_file_err))
+		return
 	}
 
 	return true, string(token_data.Login), nil
@@ -105,36 +116,42 @@ func deleteExpiredTokens() {
 	if checkFileExistance("./save_data/account_tokens/.running_deletion") {
 		return
 	}
-	_, running_file_creation_err := os.Create("./save_data/account_tokens/.running_deletion")
+	running_deletion_file, running_file_creation_err := os.Create("./save_data/account_tokens/.running_deletion")
 	if running_file_creation_err != nil {
-		serverLog("utilities.go:110", fmt.Sprintf("failed to create the file \".running_deletion\", the error \"%s\"", running_file_creation_err))
+		serverLog("utilities.go:121", fmt.Sprintf("failed to create the file \".running_deletion\", the error \"%s\"", running_file_creation_err))
 		return
 	}
+	running_deletion_file.Close()
 
 	files, _ := os.ReadDir("./save_data/account_tokens")
 	for _, file := range files {
+		if file.Name() == ".running_deletion" {
+			continue
+		}
 		token_data := AccountToken{}
 		token_file, token_err := fileToByte(fmt.Sprintf("./save_data/account_tokens/%s", file.Name()))
 		if token_err != nil {
-			serverLog("utilities.go:119", fmt.Sprintf("failed to read the token file \"%s\", the error \"%s\"", file.Name(), token_err))
+			serverLog("utilities.go:134", fmt.Sprintf("failed to read the token file \"%s\", the error \"%s\"", file.Name(), token_err))
 			continue
 		}
 		if json.Unmarshal(token_file, &token_data) != nil {
-			serverLog("utilities.go:123", fmt.Sprintf("failed to unmarshal the token file \"%s\"", file.Name()))
+			serverLog("utilities.go:138", fmt.Sprintf("failed to unmarshal the token file \"%s\"", file.Name()))
 			continue
 		}
 
 		if time.Now().Unix() > token_data.Time {
 			removal_err := os.Remove(fmt.Sprintf("./save_data/account_tokens/%s", file.Name()))
 			if removal_err != nil {
-				serverLog("utilities.go:130", fmt.Sprintf("failed to remove the token file \"%s\", the error \"%s\"", file.Name(), removal_err))
+				serverLog("utilities.go:145", fmt.Sprintf("failed to remove the token file \"%s\", the error \"%s\"", file.Name(), removal_err))
 				continue
 			}
 		}
 	}
 
+	time.Sleep(time.Second * 10)
+
 	running_file_removal_err := os.Remove("./save_data/account_tokens/.running_deletion")
 	if running_file_removal_err != nil {
-		serverLog("utilities.go:138", fmt.Sprintf("failed to remove the file \".running_deletion\", the error \"%s\"", running_file_removal_err))
+		serverLog("utilities.go:155", fmt.Sprintf("failed to remove the file \".running_deletion\", the error \"%s\"", running_file_removal_err))
 	}
 }
